Agri Data Processing Pipeline/requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Text processing and similarity
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
rapidfuzz>=3.0.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

torch>=2.0.0

anthropic>=0.72.1
pandas>=2.0.0
pytz>=2024.1
nest_asyncio>=1.6.0
